requirements.txt
Flask==2.2.3
Flask-RESTful
Flask-SQLAlchemy
Jinja2
Flask-JWT-Extended
SQLAlchemy
pymysql
Werkzeug==2.3.7
python-dotenv
marshmallow
Flask-Bcrypt
redis
celery
requests
newrelic
sib-api-v3-sdk
pymongo
Flask-PyMongo
Flask-Elasticsearch==0.2.5
py-healthcheck